// frontend/src/charting_library/charting_library.d.ts
export declare type Nominal<T, Name extends string> = T & {
  /* eslint-disable-next-line jsdoc/require-jsdoc */ [Symbol.species]: Name
}
// ...
export interface Bar {
  /** Bar time.
   * Amount of **milliseconds** since Unix epoch start in **UTC** timezone.
   * `time` for daily, weekly, and monthly bars is expected to be a trading day (not session start day) at 00:00 UTC.
   * The library adjusts time according to `session` from {@link LibrarySymbolInfo}.
   */
  time: number
  /** Opening price */
  open: number
  /** High price */
  high: number
  /** Low price */
  low: number
  /** Closing price */
  close: number
  /** Trading Volume */
  volume?: number
}
// ...
export interface HistoryMetadata {
  /**
   * Optional value indicating to the library that there is no more data on the server.
   */
  noData?: boolean
  /**
   * Time of the next bar if there is no data in the requested period.
   * It should be set if the requested period represents a gap in the data. Hence there is available data prior to the requested period.
   */
  nextTime?: number | null
}
// ...
export interface IExternalDatafeed {
  /**
   * This call is intended to provide the object filled with the configuration data.
   * The lib assumes that you will call the callback function and pass your datafeed {@link DatafeedConfiguration} as an argument.
   *
   * @param  {OnReadyCallback} callback - callback to return your datafeed configuration ({@link DatafeedConfiguration}) to the library.
   */
  onReady(callback: OnReadyCallback): void
}
// ...
export interface PeriodParams {
  /**
   * Unix timestamp (leftmost requested bar)
   */
  from: number
  /**
   * Unix timestamp (rightmost requested bar - not inclusive)
   */
  to: number
  /**
   * The exact amount of bars to load, should be considered a higher priority than `from` if your datafeed supports it
   */
  countBack: number
  /**
   * Used to identify if it's the first call of getBars
   */
  firstDataRequest: boolean
}
// ...
export type HistoryCallback = (bars: Bar[], meta?: HistoryMetadata) => void
/** Item variants within a context menu */
export type IActionVariant = IAction | ISeparator
export type IBarArray = [number, number, number, number, number, number]
export type IBasicDataFeed = IDatafeedChartApi & IExternalDatafeed
// ...
export type OnReadyCallback = (configuration: DatafeedConfiguration) => void
// ...
export type ResolutionString = Nominal<string, 'ResolutionString'>
export type ResolveCallback = (symbolInfo: LibrarySymbolInfo) => void